refactor main file into streamsync and main
